change util location
